Refactor table setup and settingS
