+++ frontend/src/panels/components/UserInvitationItem.vue
